ARC-3608: pull rte into its own component and make it memoized to avoid re-renders on polling messages (clearing values like hyperlinks)
