+++ src/modules/account/components/MaterialRequestDetailBlade/MaterialRequestConversation.tsx
@@ -1,7 +1,6 @@
 import { useGetMaterialRequestConversationInfinite } from '@account/components/MaterialRequestDetailBlade/hooks/useGetMaterialRequestConversationInfinite';
-import { useSendMaterialRequestMessage } from '@account/components/MaterialRequestDetailBlade/hooks/useSendMaterialRequestMessage';
+import { MaterialRequestConversationInput } from '@account/components/MaterialRequestDetailBlade/MaterialRequestConversationInput';
 import { MaterialRequestConversationMessage } from '@account/components/MaterialRequestDetailBlade/MaterialRequestConversationMessage';
-import { isMaterialRequestClosed } from '@account/utils/is-material-request-closed';
 import { selectCommonUser } from '@auth/store/user';
 import {
 	type MaterialRequest,
@@ -9,33 +8,14 @@ import {
 	MaterialRequestEventType,
 	MaterialRequestStatus,
 } from '@material-requests/types';
-import {
-	Button,
-	keysEnter,
-	RichTextEditorWithInternalState,
-	TagList,
-} from '@meemoo/react-components';
-import { Icon } from '@shared/components/Icon';
-import { IconNamesLight } from '@shared/components/Icon/Icon.enums';
 import { Loading } from '@shared/components/Loading';
-import { tHtml, tText } from '@shared/helpers/translate';
-import { toastService } from '@shared/services/toast-service';
+import { tHtml } from '@shared/helpers/translate';
 import type { QueryObserverResult } from '@tanstack/react-query';
 import clsx from 'clsx';
 import { noop } from 'lodash-es';
-import React, {
-	type FC,
-	useCallback,
-	useEffect,
-	useLayoutEffect,
-	useMemo,
-	useRef,
-	useState,
-} from 'react';
+import React, { type FC, useCallback, useEffect, useLayoutEffect, useRef, useState } from 'react';
 import { useSelector } from 'react-redux';
-import { v4 as uuid } from 'uuid';
 import styles from './MaterialRequestConversation.module.scss';
-import { MessageFileUpload } from './MessageFileUpload';
 
 const MATERIAL_REQUEST_CONVERSATION_PAGE_SIZE = 20;
 
@@ -54,16 +34,11 @@ export const MaterialRequestConversation: FC<MaterialRequestConversationProps> =
 }) => {
 	const scrollableRef = useRef<HTMLDivElement>(null);
 	const scrollTriggerRef = useRef<HTMLDivElement>(null);
-	const fileListRef = useRef<HTMLDivElement>(null);
 	const [hasScrolledToBottom, setHasScrolledToBottom] = useState(false);
 	const [hasNotified, setHasNotified] = useState(false);
 	const previousScrollHeightRef = useRef<number | null>(null);
 	const user = useSelector(selectCommonUser);
-	const [editorKey, setEditorKey] = useState(uuid()); // To force rich text editor to rerender
-	const editorId = `material-request-conversation--${editorKey}`;
 
-	const [currentMessage, setCurrentMessage] = useState<string>('');
-	const [selectedFiles, setSelectedFiles] = useState<File[]>([]);
 	const [fileListHeight, setFileListHeight] = useState<number>(0);
 
 	const {
@@ -78,59 +53,6 @@ export const MaterialRequestConversation: FC<MaterialRequestConversationProps> =
 		materialRequest.id,
 		MATERIAL_REQUEST_CONVERSATION_PAGE_SIZE
 	);
-
-	const { mutate: sendMessage, isPending: isSending } = useSendMaterialRequestMessage(
-		materialRequest.id
-	);
-
-	const isMessageEmpty = useMemo(() => {
-		const textContent = currentMessage.replace(/<[^>]*>/g, '').trim();
-		return !textContent;
-	}, [currentMessage]);
-
-	const sendMessageDisabled = useMemo(
-		() => isMessageEmpty || isSending || isMaterialRequestClosed(materialRequest),
-		[isMessageEmpty, isSending, materialRequest]
-	);
-
-	const handleSendMessage = useCallback(() => {
-		if (sendMessageDisabled) {
-			return;
-		}
-
-		// Send single message with text and all selected files
-		sendMessage(
-			{ message: currentMessage, files: selectedFiles },
-			{
-				onSuccess: () => {
-					setCurrentMessage('');
-					setSelectedFiles([]);
-					setEditorKey(uuid()); // Force rerender of rich text editor
-				},
-				onError: (err) => {
-					console.error(err);
-					toastService.notify({
-						maxLines: 3,
-						title: tText(
-							'modules/account/components/material-request-detail-blade/material-request-conversation___er-ging-iets-mis'
-						),
-						description: tText(
-							'modules/account/components/material-request-detail-blade/material-request-conversation___het-bericht-kon-niet-worden-verzonden'
-						),
-					});
-				},
-			}
-		);
-	}, [sendMessageDisabled, currentMessage, selectedFiles, sendMessage]);
-
-	// Measure the file list height and adjust message wrapper accordingly
-	// biome-ignore lint/correctness/useExhaustiveDependencies: We want to re-measure when files are added or removed
-	useLayoutEffect(() => {
-		if (fileListRef.current) {
-			const height = fileListRef.current.offsetHeight;
-			setFileListHeight(height);
-		}
-	}, [selectedFiles]);
 
 	// Refetch the messages when the material request gets closed while viewing the conversation to get the latest messages and reflect the closed status in the UI
 	useEffect(() => {
@@ -266,51 +188,6 @@ export const MaterialRequestConversation: FC<MaterialRequestConversationProps> =
 		return () => observer.disconnect();
 	}, [handleLoadMore, hasScrolledToBottom]);
 
-	const getEditorAndFocusOrDisable = useCallback(() => {
-		const braftComponent = document.querySelector(
-			`#${editorId} .DraftEditor-editorContainer .public-DraftEditor-content`
-		) as HTMLDivElement;
-
-		if (braftComponent) {
-			braftComponent?.focus();
-
-			const isDisabled = isMaterialRequestClosed(materialRequest);
-			const braftControls = document.querySelectorAll<HTMLElement>(
-				'.bf-controlbar button.control-item, .bf-controlbar .dropdown-handler'
-			);
-
-			braftControls.forEach((control) => {
-				control.tabIndex = isDisabled ? -1 : 0;
-			});
-
-			return true;
-		}
-		return false;
-	}, [editorId, materialRequest]);
-
-	useEffect(() => {
-		let success = getEditorAndFocusOrDisable();
-
-		if (success) {
-			return;
-		}
-
-		const observer = new MutationObserver((_, observer) => {
-			success = getEditorAndFocusOrDisable();
-
-			if (success) {
-				observer.disconnect();
-				return;
-			}
-		});
-
-		observer.observe(document.body, {
-			childList: true,
-			subtree: true,
-		});
-		return () => observer.disconnect();
-	}, [getEditorAndFocusOrDisable]);
-
 	const renderContent = () => {
 		if (isLoadingMessages) {
 			return (
@@ -371,92 +248,11 @@ export const MaterialRequestConversation: FC<MaterialRequestConversationProps> =
 							));
 					})}
 				</div>
-				<div className={clsx(styles['p-conversation-messages__editor'])}>
-					<div
-						ref={fileListRef}
-						className={clsx(styles['p-conversation-messages__selected-files'])}
-					>
-						<TagList
-							tags={selectedFiles.map((file, index) => ({
-								id: `${file.name}-${index}`,
-								label: (
-									<>
-										<Icon name={IconNamesLight.File} />
-										<span>{file.name}</span>
-										<span> ({Math.round((file.size / 1024 / 1024) * 100) / 100} MB)</span>
-									</>
-								),
-							}))}
-							closeIcon={<Icon name={IconNamesLight.Times} aria-hidden />}
-							onTagClosed={(id) => {
-								const index = Number.parseInt(String(id).split('-').pop() || '0', 10);
-								setSelectedFiles(selectedFiles.filter((_, i) => i !== index));
-							}}
-							variants={['closable', 'silver']}
-						/>
-					</div>
-					<RichTextEditorWithInternalState
-						braft={{
-							contentStyle: {
-								minHeight: '100px',
-								maxHeight: '150px',
-								overflowY: 'auto',
-							},
-							// @ts-expect-error: This method does exists on the braft editor so ts-ignoring this to get the error gone
-							keyBindingFn: (evt: KeyboardEvent) => {
-								if (evt.ctrlKey || evt.shiftKey || evt.altKey || evt.metaKey) {
-									// In case of these buttons being pressed, we will allow the enter to go through
-									// Otherwise it will be impossible to get more than 1 entry in a list
-									return;
-								}
-
-								if (keysEnter.includes(evt.key)) {
-									handleSendMessage();
-								}
-							},
-						}}
-						disabled={isMaterialRequestClosed(materialRequest)}
-						className={
-							isMaterialRequestClosed(materialRequest)
-								? styles['p-conversation-messages__editor--disabled']
-								: undefined
-						}
-						id={editorId}
-						value={currentMessage}
-						onChange={(value) => setCurrentMessage(value)}
-						placeholder={tText(
-							'modules/account/components/material-request-detail-blade/material-request-detail-blade___typ-je-bericht'
-						)}
-						controls={[
-							'bold',
-							'italic',
-							'underline',
-							'list-ul',
-							'list-ol',
-							'link',
-							{
-								type: 'customButton',
-								component: (
-									<MessageFileUpload
-										onFileSelected={(file) => setSelectedFiles((prev) => [...prev, file])}
-										disabled={isMaterialRequestClosed(materialRequest)}
-									/>
-								),
-							},
-						]}
-						key={editorKey}
-					/>
-					<Button
-						id="material-request-conversation__send-button"
-						className={clsx(styles['p-conversation-messages__editor__send-button'])}
-						variants={['text']}
-						// Replace this icon with a send icon when Jelle and JN add the icons to the font
-						icon={<Icon name={IconNamesLight.PaperPlane} />}
-						disabled={sendMessageDisabled}
-						tabIndex={sendMessageDisabled ? -1 : undefined}
-						onClick={handleSendMessage}
-					/>
-				</div>
+				<MaterialRequestConversationInput
+					key={materialRequest.id}
+					materialRequest={materialRequest}
+					onAttachmentsChanged={setFileListHeight}
+				/>
 			</>
 		);
 	};
